fix(go-sdk): Fix clippy warnings and add missing gorilla/websocket dependency
- Remove unreachable Language::Go match arm in generate.rs (separate
  TypeScript and Go formatting arms)
- Remove unused has_go() from detect.rs and check_for_go() from init.rs
- Fix unnecessary u64 casts in go.rs and smoketests/lib.rs
- Move WASM helper functions before test module in smoketests/lib.rs
  to fix items_after_test_module clippy lint
- Add gorilla/websocket dependency required by merged client SDK code

+++ crates/cli/src/tasks/go.rs
@@ -196,7 +196,7 @@ fn inject_export(wasm: &[u8], new_export: Vec<u8>) -> Option<Vec<u8>> {
 
             let new_count_encoded = {
                 let mut v = Vec::new();
-                leb_encode(&mut v, new_count as u64);
+                leb_encode(&mut v, new_count);
                 v
             };
